-Переделка шаблона для статистики по "минимальным, максимальным и средним (арифметическое и средневзвешенное) УПКС по административным округам";
git-tfs-id: [http://192.168.3.86:8080/tfs/defaultcollection]$/CIPJS/KadOzenka;C63571

+++ Source/KadOzenka.Dal/ManagementDecisionSupport/Dto/StatisticalData/MinMaxAverageUPKSByAdministrativeDistrictsDto.cs
@@ -6,7 +6,7 @@ namespace KadOzenka.Dal.ManagementDecisionSupport.Dto.StatisticalData
 	{
 		public string AdditionalName { get; set; }
 		public string Name { get; set; }
-		public int ObjectsCount { get; set; }
+		public long ObjectsCount { get; set; }
 		public UpksCalcType UpksCalcType { get; set; }
 		public string PropertyType { get; set; }
 		public decimal? UpksCalcValue { get; set; }
